fix jit bug

+++ desc/objectives/_generic.py
@@ -980,9 +980,10 @@ class DeflationOperator(_Objective):
             f = self._objective.compute(params)
             # if wrapping an objective, but all things are None, make deflation do
             # nothing when multiplying f, so here we add 1 to it as it is 0 right now
-            deflation_parameter += float(
-                jnp.invert(self._not_all_things_to_deflate_are_None)
-            )
+            deflation_parameter += jnp.invert(
+                self._not_all_things_to_deflate_are_None
+            ).astype(float)
+
         else:
             f = 1.0
